Implemented player vision and moved files around
Moved files around in order to create a more organized file structure
within unity. Was also finally about to get a very rough and unoptimized
version of player vision running in the game, it uses bresenham's line
algorithm to cast rays all around the player towards all the edges of
the map. These rays are then followed, turning all floor tiles to
visible until a wall is reached, the wall is then made visible and we
move onto the next ray. This essentially gives our player a 360 field of
view but it has some flaws which I have began to address with the
PostProcessing function that is currently within the PlayerEntity.
Specifically these lines need to be antialliased, the most popular
solution online, and what I have implemented, is correcting the lines
sight according to their "section" around the player. Also the function
used to create the bresenham lines in this project was almost entirely
based off of the answer to this question https://stackoverflow.com/questions/11678693/all-cases-covered-bresenhams-line-algorithm