Added level exits
The dungeon now generates exits that as for now just reloads a new map
and increments a level var (right now this var does not factor into the
generation but this can later be used to make future levels harder).
I also moved around control of certain functions to make more logical
sense, the generation of the player spawn point is now handled by the
board class while the actual creation of the player is handled by
EntityController which simply gets the board to use from the board
controller. I've also removed the requirement for certain scripts to be
disabled and reenabled as that just seemed like a dirty workaround that
would lead to issues in the future.