Added vision limitation
Vision can now be limited with a set value, this is already applied to
the player but will also work automatically with all future entities. I
had issues getting the line algorithm to cast to a limited square (the
lines where too sharp as the area got smaller, meaning certain tiles
were entirely missed) so as of now the vision logic still calls for rays
to the edges of the map, and then simply removes any enabled tiles
outside of the range of vision after the fact.