Changed to pure white sprites, started on water
Moved the sprites we were using (which were pre colored grey) over to a
new sprite sheet which is actually white, letting us control the color
from within unity (and also making white look alot better....). Also
started on water generation for the dungeon/cave map, currently the
water is placed in the inner blank areas that are generated during the
game of life.