Worked on water generation for cave map
Added abit to water generation over the past few days. Some of the code
is currently commented out because I was in the process of writing it
but am giving a presentation over Cellular Automata generation today and
wanted the current version of the game to be able to show during the
presentation.